Result of passing FindBugs.
git-svn-id: svn://localhost/trunk@1859 7f4d9e04-a92a-ab41-bea9-970b690ef4a7

+++ org.python.pydev/src/org/python/pydev/editor/PyInformationPresenter.java
@@ -104,7 +104,7 @@ public class PyInformationPresenter implements DefaultInformationControl.IInform
         try {
 
             StringBuffer buffer= new StringBuffer();
-            int maxNumberOfLines= Math.round(maxHeight / gc.getFontMetrics().getHeight());
+            int maxNumberOfLines= Math.round((float)maxHeight / (float)gc.getFontMetrics().getHeight());
 
             fCounter= 0;
             PyLineBreakReader reader= new PyLineBreakReader(createReader(hoverInfo, presentation), gc, maxWidth);
